backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
python-multipart==0.0.9
pdfplumber==0.11.0
pypdf==4.2.0
pytesseract==0.3.13
pdf2image==1.17.0
python-docx==1.1.2
scikit-learn==1.4.2
numpy<2
python-dotenv==1.0.1
httpx==0.27.0
email-validator==2.1.1
